refactor(auth): rename folder
