+++ src/modules/auth/SignUp/services/types/fetchPostSignUpService.types.ts
